+++ client/store/singlePackStore.js
@@ -1,16 +1,18 @@
 // redux toolkit slice for single pack
 
-import { createSlice, createAsyncThunk, createEntityAdapter} from "@reduxjs/toolkit";
+import {
+  createSlice,
+  createAsyncThunk,
+  createEntityAdapter,
+} from "@reduxjs/toolkit";
 
 import axios from "axios";
 
 import { api } from "../constants/api";
 
-const singlePackAdapter = createEntityAdapter(
-  {
-    selectId: (singlePack) => singlePack._id,
-  }
-);
+const singlePackAdapter = createEntityAdapter({
+  selectId: (singlePack) => singlePack._id,
+});
 
 // Step 2: Create the initial state using the entity adapter
 const initialState = singlePackAdapter.getInitialState({
@@ -24,6 +26,24 @@ export const fetchSinglePack = createAsyncThunk(
   async (packId) => {
     const response = await axios.get(`${api}/pack/p/${packId}`);
     return response.data;
+  }
+);
+export const selectItemsGlobal = createAsyncThunk(
+  "Items/selectItemsGlobal",
+  async (item) => {
+    try {
+      const itemId = item.selectedItem;
+      const ownerId = item.ownerId;
+      const packId = item.packId;
+
+      const response = await axios.post(`${api}/item/global/select/${packId}`, {
+        itemId: itemId,
+        ownerId: ownerId,
+      });
+      return response.data;
+    } catch (error) {
+      console.log("error", error.message);
+    }
   }
 );
 
@@ -44,6 +64,19 @@ const singlePackSlice = createSlice({
         state.error = null;
       })
       .addCase(fetchSinglePack.rejected, (state, action) => {
+        state.isLoading = false;
+        state.error = action.error.message;
+      })
+      .addCase(selectItemsGlobal.pending, (state) => {
+        state.isLoading = true;
+        state.error = null;
+      })
+      .addCase(selectItemsGlobal.fulfilled, (state, action) => {
+        state.singlePack.items.push(action.payload.data);
+        state.isLoading = false;
+        state.error = null;
+      })
+      .addCase(selectItemsGlobal.rejected, (state, action) => {
         state.isLoading = false;
         state.error = action.error.message;
       });
